requirements.txt
torch==2.4.0+cpu
torchvision==0.19.0+cpu
numpy<2
Pillow
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
redis
--extra-index-url https://download.pytorch.org/whl/cpu
